Added ImmutablePlatformModule class, that is a superclass of PlatformModule and is the one passed to all inputs and outputs. Also, created a common MatchMakingLoggableElement abstract class so it can be extended and therefore have access to the debug input everywhere
