+++ src/main/java/InfrastructureManager/Modules/MatchMaking/Input/MatchMakerInput.java
@@ -1,6 +1,7 @@
 package InfrastructureManager.Modules.MatchMaking.Input;
 
 import InfrastructureManager.ModuleManagement.Exception.Execution.ModuleExecutionException;
+import InfrastructureManager.ModuleManagement.ImmutablePlatformModule;
 import InfrastructureManager.ModuleManagement.ModuleInput;
 import InfrastructureManager.ModuleManagement.PlatformModule;
 import InfrastructureManager.Modules.MatchMaking.MatchesList;
@@ -10,7 +11,7 @@ public class MatchMakerInput extends ModuleInput {
     private final MatchesList sharedMatchesList;
     private String toSend;
 
-    public MatchMakerInput(PlatformModule module, String name, MatchesList mapping) {
+    public MatchMakerInput(ImmutablePlatformModule module, String name, MatchesList mapping) {
         super(module,name);
         toSend = "";
         this.sharedMatchesList = mapping;
