+++ src/main/java/InfrastructureManager/Modules/NetworkStructure/Input/NetworkInput.java
@@ -1,6 +1,7 @@
 package InfrastructureManager.Modules.NetworkStructure.Input;
 
 import InfrastructureManager.ModuleManagement.Exception.Execution.ModuleExecutionException;
+import InfrastructureManager.ModuleManagement.ImmutablePlatformModule;
 import InfrastructureManager.Modules.NetworkStructure.Exception.NetworkModuleException;
 
 import com.fasterxml.jackson.core.JsonProcessingException;
@@ -13,7 +14,7 @@ public class NetworkInput extends ModuleInput {
 	
 	private final Network network;
 	
-	public NetworkInput(PlatformModule module,String name,Network network) {
+	public NetworkInput(ImmutablePlatformModule module, String name, Network network) {
 		super(module,name);
 		this.network = network;
 	}
